refactor: move onboarding components
