+++ mobile/src/styles/onboarding.styl
@@ -14,6 +14,25 @@
     .wizard-button
         margin-bottom 30px
 
+    .wizard-button--with-circles
+        margin-bottom 10px
+
+    .wizard-circles
+        display flex
+        justify-content center
+        margin-bottom 5px
+
+    .wizard-circles__circle
+        width 15px
+        height 15px
+        margin 0 5px
+        background rgba(0,0,0,.3)
+        border-radius 50px
+        transition all .3s ease
+
+    .wizard-circles__circle--active
+        bacground #297ef2
+
 .welcome
     background-color white
     color            black
